🤖 Refactor: parser에 검증 로직 제거
