+++ packages/terraform/parser/LaunchConfigurationParser.ts
@@ -1,15 +1,9 @@
 import { BaseResourceParser } from './BaseResourceParser';
 import { ValidationError } from '../util/ValidationError';
-import { NCloudLaunchConfiguration } from '../model/NCloudLaunchConfiguration';
+import { NCloudLaunchConfiguration } from '../model/autoScaling/NCloudLaunchConfiguration';
 
 export class LaunchConfigurationParser extends BaseResourceParser {
     protected resourceType = ['launchconfiguration'];
-
-    protected validateProperties(properties: any): void {
-        if (!properties.name) {
-            throw new ValidationError('LaunchConfiguration', 'name', 'name이 필수입니다');
-        }
-    }
 
     protected createModel(properties: any): NCloudLaunchConfiguration {
         return new NCloudLaunchConfiguration({
@@ -19,7 +13,7 @@ export class LaunchConfigurationParser extends BaseResourceParser {
             memberServerImageNo: properties.memberServerImageNo,
             loginKeyName: properties.loginKeyName,
             initScriptNo: properties.initScriptNo,
-            isEncryptedVolume: properties.isEncryptedVolume
+            isEncryptedVolume: properties.isEncryptedVolume,
         });
     }
 }
