--- packages/terraform/parser/interface/MySQLValidationRules.ts
@@ -1,6 +1,0 @@
-interface MySQLValidationRules {
-    fieldName: string,
-    minLength?: number,
-    maxLength?: number,
-    required: boolean;
-}
